+++ src/main/java/com/thxpapa/juneberrydiary/repository/fileRepository/JuneberryFileRepository.java
@@ -3,5 +3,5 @@ package com.thxpapa.juneberrydiary.repository.fileRepository;
 import com.thxpapa.juneberrydiary.domain.file.JuneberryFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JuneberryFileRepository extends JpaRepository<JuneberryFile, Integer> {
+public interface JuneberryFileRepository extends JpaRepository<JuneberryFile, Long> {
 }
