+++ src/main/java/com/thxpapa/juneberrydiary/repository/geoRepository/SpotRepository.java
@@ -3,6 +3,6 @@ package com.thxpapa.juneberrydiary.repository.geoRepository;
 import com.thxpapa.juneberrydiary.domain.geo.Spot;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpotRepository extends JpaRepository<Spot, Integer> {
+public interface SpotRepository extends JpaRepository<Spot, Long> {
 
 }
