+++ src/main/java/com/thxpapa/juneberrydiary/repository/postRepository/PostFileRepository.java
@@ -3,5 +3,5 @@ package com.thxpapa.juneberrydiary.repository.postRepository;
 import com.thxpapa.juneberrydiary.domain.post.PostFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostFileRepository extends JpaRepository<PostFile, Integer> {
+public interface PostFileRepository extends JpaRepository<PostFile, Long> {
 }
